requirements.txt
bottleneck
tensorflow-gpu==2.6
keras==2.6
numpy
matplotlib
opencv-python
opencv-contrib-python
scipy
scikit-learn
scikit-image
Pillow
tqdm
tensorboard